workplace/server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0
openai>=1.50.0
httpx>=0.27.0
